refactor: rename signature parser file
