+++ src/suricatals/langserver.py
@@ -30,7 +30,7 @@ from typing import Optional
 from importlib.metadata import version
 import queue
 
-from suricatals.parse_signatures import SuricataFile
+from suricatals.signature_parser import SuricataFile
 from suricatals.signature_validator import TestRules
 from suricatals.suri_cmd import SuriCmd
 from suricatals.tokenize_sig import SuricataSemanticTokenParser
